+++ tests/testthat/test-connect.R
@@ -151,8 +151,3 @@ test_that("Add logs to connectors object", {
     expect_s3_class(x, "ConnectorLogger")
   })
 })
-
-test_that("Metadata works correctly", {
-  # Create connectors with metadata
-  cnts <- connect(yaml_file)
-})